# Parquetquery join performance improvement