@mdisibio 2024-05-06
> i was also wondering what happens if we force seek all iters once and then do an initial sort by distance into the file.

Yes, I did experiment with that and sorting once after first peek (no dynamic reordering) wasn't beneficial.  It was either neutral or slightly worse depending on the query.  I think basically it's too brittle and has a high chance of getting locked into a poor pattern from the first matches.   Combining both (sort on first peek and in the loop) didn't help but also logically it wasn't expected to since the initial sort is immediately overwritten.